tests: move verify_fs_exist to tests.helpers
